requirements.txt
# AI Suspect Sketch Generator v2.1 — Dependencies
# Install: pip install -r requirements.txt

# Core
torch>=2.2.0
torchvision>=0.17.0

# Image generation backends (API-based — no local model download needed)
requests>=2.31.0
Pillow>=10.2.0

# LLM — Groq (Llama-3)
groq>=0.5.0

# Face validation (optional but recommended)
facenet-pytorch>=2.5.3

# Environment
python-dotenv>=1.0.0

# API backend
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# Frontend
streamlit>=1.32.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0

# Utilities
numpy>=1.26.0
tqdm>=4.66.0

diffusers>=0.37.0
transformers>=5.0.0
accelerate>=1.0.0
peft>=0.15.0
safetensors>=0.4.0
